Update Servlet classes for role base admin login

+++ src/main/java/com/tutor/servlet/DeleteRecordServlet.java
@@ -1,5 +1,6 @@
 package com.tutor.servlet;
 
+import com.tutor.model.Admin;
 import com.tutor.util.DataStorage;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -14,12 +15,19 @@ public class DeleteRecordServlet extends HttpServlet {
 
     @Override
     public void init() throws ServletException {
-        String filePath = "F:\\sliit\\1st year\\2 sem\\Object Oriented Programming - SE1020\\project\\sample\\HTB_Admin\\src\\main\\webapp\\records.txt";
+        String filePath = "F:\\sliit\\1st year\\2 sem\\Object Oriented Programming - SE1020\\project\\Final real project\\HTBT-Admin CRUD\\src\\main\\webapp\\records.txt";
         this.dataStorage = new DataStorage(filePath);
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+
+        Admin user = (Admin) request.getSession().getAttribute("user");
+        if (user == null || !"Admin".equals(user.getRole())) {
+            response.sendRedirect(request.getContextPath() + "/list");
+            return;
+        }
+
         int id = Integer.parseInt(request.getParameter("id"));
         try {
             dataStorage.deleteRecord(id);
